+++ website/react-magma-docs/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
+## 3.0.0-next.24
+
+### Patch Changes
+
+- Updated dependencies [ce9088243]
+  - react-magma-dom@3.0.0-next.17
+
+## 3.0.0-next.23
+
+### Major Changes
+
+- 50d3b20ee: homepage: updating home page with new design and rebrand colors
+
 ## 3.0.0-next.22
 
 ### Major Changes